+++ runtime/nonfb/FbtEnv.js
@@ -1,0 +1,16 @@
+/**
+ * Copyright 2004-present Facebook. All Rights Reserved.
+ *
+ * @emails oncall+internationalization
+ * @flow strict
+ * @format
+ */
+
+/**
+ * An empty no-op for now.
+ */
+const FbtEnv = {
+  setup() {},
+};
+
+module.exports = FbtEnv;